requirements.txt
pillow==11.1.0
python-dotenv==1.0.1
qrcode==8.0
